backend/requirements-label.txt
brother-ql==0.9.4
Pillow==11.1.0
pyusb==1.3.1
